MovePdfsAPI/Py-scripts/requirements.txt
beautifulsoup4
mysql-connector-python==8.0.28
nltk
rake-nltk